Move helper functions from sentis-kernel-metadata to sentis.js
